+++ private/sys/Controller/Public/PageController.php
@@ -677,7 +677,8 @@ final class PageController
             $this->pageBlocks = new PageBlocks(
                 dirname(__DIR__, 4),
                 $this->pageBlockParser(),
-                $this->pageMarkdown()
+                $this->pageMarkdown(),
+                $this->extensionServicesProvider
             );
         }